+++ dook/core/events/errors.py
@@ -1,0 +1,1 @@
+EVENT_SUBSCRIBER_NOT_FOUND_ERROR = "Event subscriber not found <{}>"
