+++ dook/settings/__init__.py
@@ -1,15 +1,3 @@
-"""
-Django settings for sfnf-panel project.
-
-Generated by "django-admin startproject" using Django 3.0.4.
-
-For more information on this file, see
-https://docs.djangoproject.com/en/3.0/topics/settings/
-
-For the full list of settings and their values, see
-https://docs.djangoproject.com/en/3.0/ref/settings/
-"""
-
 import os
 import pathlib
 import sys
@@ -49,7 +37,7 @@ LOCAL_APPS = [
     "dook.users",
     "dook.news",
     "dook.core.processor",
-    "dook.events",
+    "dook.core.events",
 ]
 
 THIRD_PARTY_APPS = [
